[test] 결제 로직 테스트 수정
issue: 88

+++ src/main/java/com/mini/joymall/payment/domain/repository/PaymentRepository.java
@@ -3,6 +3,6 @@ package com.mini.joymall.payment.domain.repository;
 import com.mini.joymall.payment.domain.entity.Payment;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PaymentRepository extends CrudRepository<Payment, Integer> {
+public interface PaymentRepository extends CrudRepository<Payment, Long> {
 
 }
